-ActionTabsPanel change in JLabel string -Moved JLoginPanel, JRegisterPanel -GUI_view modificatiosn to imports, adding model, changing continue access after login -removed impports due to moving JLogin and JRegister from StartUpPanel
