--- src/main/java/GUI/GUI_Done/JRegisterFrame.java
@@ -1,19 +1,0 @@
-package GUI.GUI_Done;
-
-import javax.swing.*;
-
-public class JRegisterFrame extends JFrame {
-    public JRegisterFrame(){
-        super("Register Window");
-        JRegisterPanel registerPanel = new JRegisterPanel();
-
-        this.add(registerPanel);
-        pack();
-        setLocationRelativeTo(null);
-        setVisible(true);
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-    }
-    public static void main(String [] args){
-        JRegisterFrame frame = new JRegisterFrame();
-    }
-}
